finalizing scripts
